python/pyproject.toml
[project]
name = "anki-builder"
version = "0.1.0"
description = "Helper which manages anki decks based on pre-parsed CSV files"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "genanki>=0.13.1",
    "pillow>=11.3.0",
    "requests>=2.32.5",
]
